Refactor entity access into custom hooks
